requirements.txt
RPi.GPIO==0.7.0
paho-mqtt==1.5.0
